+++ packages/datasource-sequelize/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/datasource-sequelize [1.5.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sequelize@1.5.4...@forestadmin/datasource-sequelize@1.5.5) (2023-08-31)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.13.0
+
 ## @forestadmin/datasource-sequelize [1.5.4](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sequelize@1.5.3...@forestadmin/datasource-sequelize@1.5.4) (2023-08-30)